+++ pyplejd/__init__.py
@@ -50,6 +50,7 @@ class PlejdManager:
         self.options = {}
         self.connection_monitor = None
         self.button_events_enabled = True
+        self.route_control_writes_directly = False
         self.reconnect_after_control_write = False
         self.control_confirmation_timeout = 1.5
         self.reauthenticate_before_reconnect = True